+++ tests/plugins/table/headers2row1col.js
@@ -13,15 +13,6 @@
 	};
 
 	bender.test( {
-		_should: {
-			ignore: {
-				// (#2881), (#2996).
-				'2 rows, 1 col, none -> both': true,
-				'2 rows, 1 col, row -> both': true,
-				'2 rows, 1 col, both -> col': true
-			}
-		},
-
 		'2 rows, 1 col, none -> none': testHeadersManipulation( 'none', 'none' ),
 
 		'2 rows, 1 col, none -> col': testHeadersManipulation( 'none', 'col' ),